refactor: move annotation folder
