+++ user-service/src/main/java/kr/mafoo/user/annotation/RequestMemberId.java
@@ -1,4 +1,4 @@
-package kr.mafoo.user.config;
+package kr.mafoo.user.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
